web_Appv3/requirements.txt
Flask==3.0.0
firebase-admin==6.5.0
pyrebase4==4.7.1
flask-cors==4.0.0
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
geopy==2.4.1
gunicorn==22.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1